fixed the spheres and the textures

+++ t1/MyTriangle.js
@@ -19,17 +19,17 @@ MyTriangle.prototype.initBuffers = function () {
 	console.log(this.args);
   //point1
   var args=this.args;
-    var x1 = parseInt(args[0]);
-    var y1 = parseInt(args[1]);
-    var z1 = parseInt(args[2]);
+    var x1 = parseFloat(args[0]);
+    var y1 = parseFloat(args[1]);
+    var z1 = parseFloat(args[2]);
 //point2
-    var x2 = parseInt(args[3]);
-    var y2 = parseInt(args[4]);
-    var z2 = parseInt(args[5]);
+    var x2 = parseFloat(args[3]);
+    var y2 = parseFloat(args[4]);
+    var z2 = parseFloat(args[5]);
 //point3
-    var x3 = parseInt(args[6]);
-    var y3 = parseInt(args[7]);
-    var z3 = parseInt(args[8]);
+    var x3 = parseFloat(args[6]);
+    var y3 = parseFloat(args[7]);
+    var z3 = parseFloat(args[8]);
 
 
     this.vertices = [
